requirements.txt
ansicon==1.89.0
arrow==1.2.3
asgiref==3.7.2
blessed==1.20.0
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.4
django-cors-headers==4.2.0
django-filter==23.2
django-picklefield==3.1
django-q==1.3.9
djangorestframework==3.14.0
drf-yasg==1.21.7
gunicorn==21.2.0
idna==3.4
inflection==0.5.1
jinxed==1.2.0
packaging==23.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
redis==3.5.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
wcwidth==0.2.6